+++ docs/source/CHANGELOG.rst
@@ -6,6 +6,7 @@ This document records notable changes to `~ <https://github.com/QSD-Group/QSDsan
 
 Ongoing
 -------
+- Added an :class:`Equipment` class for design and costing of unit equipment.
 - Added Kendall's tau test to :func:`qsdsan.stats.get_correlation`.